+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CommonOPF Changelog
 
+## dev
+-  `delete!(p.phases_into_bus, j)` in `delete_bus_j!`
+
 ## v0.2.2
 - add `MultiPhaseVariableContainerType`
 - add `reg_busses`, `turn_ratio`, `has_vreg`, `vreg`